pyproject.toml
[project]
name = "data-science"
version = "0.1.0"
description = "Pipeline Asset Management ADK Agent for lifecycle analysis and risk assessment"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Your Name", email = "your.email@company.com"},
]

dependencies = [
    "python-dotenv>=1.0.1",
    "google-adk>=1.8.0",  # ← UPDATE to latest version
    "immutabledict>=4.2.1",
    "sqlglot>=26.10.1",
    "db-dtypes>=1.4.2",
    "regex>=2024.11.6",
    "tabulate>=0.9.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-cloud-bigquery>=3.11.0",         # ← For BigQuery vector search
    "google-cloud-bigquery-storage>=2.24.0", # ← For fast streaming reads
    "google-cloud-storage>=2.10.0",          # ← For GCS document download
    "PyPDF2>=3.0.0",                         # ← For PDF text extraction
    "python-docx>=1.1.0",                    # ← For Word document extraction
    "absl-py>=2.2.2",
    "pydantic>=2.11.3",
    "pandas>=2.3.0",
    "numpy>=2.3.1",
    "pg8000>=1.31.2",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",    
    "weasyprint>=60.0", 
    "pygments>=2.17.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.5.0",
    "nest-asyncio>=1.6.0",  # ← ADD THIS
]

jupyter = [
    "jupyter~=1.0.0",
]

lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
]

[build-system]
requires = ["hatchling"]  # ← CHANGE from uv_build
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ========== LINTING CONFIGURATION ==========
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ========== TYPE CHECKING ==========
[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

# ========== SPELL CHECKING ==========
[tool.codespell]
skip = "./locust_env/*,uv.lock,.venv,**/*.ipynb"

# ========== TESTING ==========
[tool.pytest.ini_options]
pythonpath = "."
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::UserWarning",
]

# ========== AGENT STARTER PACK ==========
[tool.agent-starter-pack]
base_template = "adk_base"
description = "Pipeline Asset Management Agent for infrastructure lifecycle analysis, risk assessment, and capital planning"
example_question = "Which pipes should we replace in the next 5 years?"

[tool.agent-starter-pack.settings]
deployment_targets = ["agent_engine", "cloud_run"]
interactive_command = "dev"